Moved common Kafka and Mqqt files
