+++ .idea/runConfigurations/connect_adapters_iiot.xml
@@ -1,15 +1,16 @@
 <component name="ProjectRunConfigurationManager">
-    <configuration default="false" name="connect-adapters-iiot" type="Application" factoryName="Application">
-        <envs>
-            <env name="SP_PORT" value="8002" />
-            <env name="SP_DATA_LOCATION" value="./test_data/" />
-            <env name="SP_DEBUG" value="true" />
-            <env name="SP_DEBUG_CONNECT" value="false" />
-        </envs>
-        <option name="MAIN_CLASS_NAME" value="org.apache.streampipes.connect.iiot.ConnectAdapterIiotInit" />
-        <module name="streampipes-connect-adapters-iiot" />
-        <method v="2">
-            <option name="Make" enabled="true" />
-        </method>
-    </configuration>
+  <configuration default="false" name="connect-adapters-iiot" type="Application" factoryName="Application">
+    <envs>
+      <env name="SP_PORT" value="8002" />
+      <env name="SP_DATA_LOCATION" value="./test_data/" />
+      <env name="SP_DEBUG" value="true" />
+      <env name="SP_DEBUG_CONNECT" value="false" />
+      <env name="SP_HOST" value="172.31.0.10" />
+    </envs>
+    <option name="MAIN_CLASS_NAME" value="org.apache.streampipes.connect.iiot.ConnectAdapterIiotInit" />
+    <module name="streampipes-connect-adapters-iiot" />
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
+  </configuration>
 </component>
