+++ streampipes-connect-adapters-iiot/pom.xml
@@ -37,6 +37,11 @@
             <artifactId>streampipes-connect-container-worker</artifactId>
             <version>${streampipes.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-pipeline-elements-shared</artifactId>
+            <version>${streampipes.version}</version>
+        </dependency>
 
         <!-- External dependencies -->
         <dependency>
