+++ streampipes-connect-adapters/pom.xml
@@ -37,6 +37,11 @@
             </dependency>
             <dependency>
                 <groupId>org.apache.streampipes</groupId>
+                <artifactId>streampipes-pipeline-elements-shared</artifactId>
+                <version>${streampipes.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.streampipes</groupId>
                 <artifactId>streampipes-container-extensions</artifactId>
                 <version>${streampipes.version}</version>
             </dependency>
