+++ streampipes-connect-adapters/src/main/java/org/apache/streampipes/connect/adapters/flic/FlicMQTTAdapter.java
@@ -19,13 +19,11 @@
 package org.apache.streampipes.connect.adapters.flic;
 
 import com.google.gson.Gson;
-import org.apache.streampipes.connect.utils.MqttConnectUtils;
 import org.apache.streampipes.connect.adapter.Adapter;
 import org.apache.streampipes.connect.api.exception.AdapterException;
 import org.apache.streampipes.connect.adapter.model.pipeline.AdapterPipeline;
 import org.apache.streampipes.connect.adapter.model.specific.SpecificDataStreamAdapter;
-import org.apache.streampipes.connect.protocol.stream.MqttConfig;
-import org.apache.streampipes.connect.protocol.stream.MqttConsumer;
+import org.apache.streampipes.pe.shared.config.Mqqt.*;
 import org.apache.streampipes.messaging.InternalEventProcessor;
 import org.apache.streampipes.model.AdapterType;
 import org.apache.streampipes.model.connect.adapter.SpecificAdapterStreamDescription;
