+++ streampipes-connect-adapters/src/main/java/org/apache/streampipes/connect/adapters/ti/TISensorTag.java
@@ -18,7 +18,7 @@
 
 package org.apache.streampipes.connect.adapters.ti;
 
-import org.apache.streampipes.connect.utils.MqttConnectUtils;
+import org.apache.streampipes.pe.shared.config.Mqqt.*;
 import org.apache.streampipes.sdk.helpers.Locales;
 import org.apache.streampipes.sdk.utils.Assets;
 import org.slf4j.Logger;
@@ -27,8 +27,6 @@ import org.apache.streampipes.connect.adapter.Adapter;
 import org.apache.streampipes.connect.api.exception.AdapterException;
 import org.apache.streampipes.connect.adapter.model.pipeline.AdapterPipeline;
 import org.apache.streampipes.connect.adapter.model.specific.SpecificDataStreamAdapter;
-import org.apache.streampipes.connect.protocol.stream.MqttConfig;
-import org.apache.streampipes.connect.protocol.stream.MqttConsumer;
 import org.apache.streampipes.messaging.InternalEventProcessor;
 import org.apache.streampipes.model.AdapterType;
 import org.apache.streampipes.model.connect.adapter.SpecificAdapterStreamDescription;
