+++ streampipes-pipeline-elements-shared/src/main/java/org/apache/streampipes/pe/shared/config/Kafka/KafkaConnectUtils.java
@@ -16,9 +16,9 @@
  *
  */
 
-package org.apache.streampipes.connect.iiot.utils;
+package org.apache.streampipes.pe.shared.config.Kafka;
 
-import org.apache.streampipes.connect.iiot.protocol.stream.KafkaConfig;
+import org.apache.streampipes.pe.shared.config.Kafka.KafkaConfig;
 import org.apache.streampipes.model.staticproperty.StaticPropertyAlternative;
 import org.apache.streampipes.sdk.StaticProperties;
 import org.apache.streampipes.sdk.extractor.StaticPropertyExtractor;
